IMS-VT: Fix surface issue when InCallActivity is not destroyed.
If any dialog is being displayed when all calls end, InCallActivity
is not destroyed. However the cached surfaces are cleaned up when
all calls end. When next VT call is initiated, new surfacetexture
available callbacks are not invoked as the framework will re use
the old resources. This will cause VT call preview and far end
views to be stuck with previous frame.
Fix this by cleaning up cached surfaces only when corresponding
SurfaceTextures are destroyed. If SurfaceTextures are not destroyed
even after all calls end, reuse them for the next call.

Change-Id: I7d7f4a64f36d81cabfeb6655e3f65bc577c6d9bc
CRs-Fixed: 2043120