IMS-VT: Close camera on previous call Video Call Object
When a Waiting VT call is accepted as VT-RX, expectation
is camera should be closed. VideoCallPresenter is instructing
to close the camera but is not honoured at lower layers since
Active Video call has opened the camera and is the camera owner.

When user tries to answer waiting VT call, existing active call
is put on hold. So, now there are two calls, one in hold state and
other in incoming state that results in primary call change.
Existing call is put on hold and new call is in incoming state
does mean that user is trying to answer the call. So, close the
camera on previous call i.e. hold call video call object.

By doing this, instructions to open/close camera on new call will
be honoured since cameraOwer will be properly set at lower layers.

Also, if user tries to open camera application immediately after
downgrading Video call to Voice call, VideoCallPresenter will
de-register itself from call state change indications. So, it will
not know of callType change to Voice from Video and will not close
camera. To fix this, close camera when VideoCallFragment stops that
happens when UE goes to background.

Change-Id: I06649715923e08faea0fe07bbc04707d25a7da19
CRs-Fixed: 2079456