Don't ignore the call notification when incall UI is shown
If the state of incoming call does not become active,
the incall UI would ignore the notification when in call UI
is shown, otherwise if the state of outgoing call does not
become active, the notification would be ignored when incall
UI goes to background.

To fix it, don't ignore the notification, and notify it as
quiet notification for incoming call when incall UI is shown.

Change-Id: Ie08c3b2291f79d43fa69c55cc306ef22f3e03509
CRs-Fixed: 2081812