IMS-VT: Refresh Video UI when UE comes back to foreground
When a VT call that is in background is pulled to foreground from HUN,
activity is getting recreated. As part of this, VideoCallFragment recreates
and Video Call UI is getting refreshed though there are no changes in the
call state/video state etc. After video UI refreshes, VideoCallPresenter
registers for indications from InCallUIStateNotifier and gets immediately
notified of whether UI is showing or not. At this point, video UI is not
refreshed again and we only set pause image based on whether UE is in
foreground or background. To fix this, refresh video UI whenever
VideoCallPresenter is notified of whether UI is showing or not.

Change-Id: I2a97a4697e0e6c73c0c01a4a0385d4e4804dab15
CRs-Fixed: 2084263