IMS-VT: Do not change camera state when already in a Video call
Do not change camera state when already in a Video call eg. when
an upgrade request is initiated or is received etc. The camera
state will be changed only after the user action on MT upgrade
request results in a call state change indication with an
appropriate video state from lower layers.

Change-Id: I0c1a65bb8a6f1d5f511d8203b083d11ae86a35d6
CRs-Fixed: 2086137