IMS-VT: Fix crash for incoming VT call on top of upgrade request.
Preview surface during incoming VT call should be handled by
AnswerFragment and the corresponding VideoCallPresenter.
However VideoCallPresenter associated with VideoCallFragment is
wrongly handling the incoming call event just after it was
torn down. This causes local preview surface to be attached to
textureview of VideoCallFragment and results in a crash due to
wrong state of surface texture.
To fix this, ignore call state events in VideoCallPresenter when
UI is not ready.

Change-Id: I2246f71c96a7a6a9a352d66b7309e3afeaa04973
CRs-Fixed: 2074401