IMS: Fix to hide More Options button when SRVCC happens in Incoming state.
1. When SRVCC happens in INCOMING state, PHONE_ID_EXTRA is removed but
UI for AnswerFragment is not updated. Hence, more options button keeps on showing.
To fix this, implement DetailsChangeListener and invoke the new API updateAnswerScreenUI().
This will inturn invoke updateUI() of AnswerFragment.
2. When SRVCC happens in ACTIVE state, if bottomsheet/related dialog is displaying already,
it is not dismissed. Ideally, bottomsheet and related dialogs should be dismissed when
more button's visibility is Gone.

Change-Id: I2eb4ec266bb8166aa94b35177556e4b081e49c34
CRs-Fixed: 2078203