Reduce binder calling times and refreshing UI
When In-call UI is brought up, lots of remote calling
are related to API 'getSystemService', which consuming
so much CPU time. Otherwise, in connecting and dialing
state, many state changed events are performed so that
frequently refresh UI, but no more info is needed to
display.

To fix it, reduce calling API 'getSystemService', and
just perform work flow once for call state DIALING
and CONNECTING separately if details does not change.

Change-Id: Iffadf85b7c10e213d608eb508f7b8cd2bd0978c5
CRs-Fixed: 2089192