IMS-VT: Prevent closing camera wrongly during back to back VT call.
InCallActivity is created when a new call is made. When all calls
are ended, InCallActivity is destroyed. If back to back MT Video
calls are received, there is a chance that InCallActivity of
previous call gets destroyed after new InCallActivity is created.
This causes camera to close while new call is on going.
To prevent this, check whether another InCallActivity is already
set before notifying onStop() of current InCallActivity.

Change-Id: Id254cedcd28c48dde358321a59d5d66a9e421ad1
CRs-Fixed: 2128620