IMS-VT: Low Battery Handling for MO and MT Video calls
When UE is under low battery:
1. User decides to place a Video call then,
   Interrupt the MO Video call and seek
   user confirmation to convert the video call to voice call
   or continue a video call
2. User receives a Video call and when user decides to answer
   the call as Video then, seek user confirmation to convert the
   video call to voice call or continue as video call
3. Disconnect the call if user presses back key with
   low battery dialog showing
4. Consume back key press for MT/Active Video call thereby not giving
   the option to dimiss low battery dialog
Change-Id: I0da978c3988d77dd0b6ed2eb9fef8d8fd2f7f43f

IMS-VT: Handle configuration changes on low battery Video call

If UE orientation changes with low battery dialog showing, then
low battery dialog is dismissed and <mAlert> is set to NULL.
This dialog is later shown to user when InCallActivity is
is recreated. At this time, onDimiss of AlertDialog is
getting called that is bringing down the dialog. Fix this
by not setting <mAlert> to NULL in onDismiss handler of
AlertDialog
Also remove code done with <onConfigurationChanging> API that
handles configuration changes on low battery dialog since this
is implicitly handled with <onDestroyInCallActivity> API.
Change-Id: I74c272b22dcb88e2e741690e2cf46af3ebf36f94

IMS-VT: Handle InCallActivity destroy when low battery dialog is showing

Handle InCallActivity destroy by dismissing the existing low battery dialog
and marking the entry against the call in low battery map that the low battery
indication needs to be reprocessed for eg. when user brings back the call to
foreground by pulling it from notification bar.
Change-Id: I0ef5912e39586a601774d7ce450daeadd73d830a

IMS-VT: Set notificationType for IncomingCall when InCallUi is not showing

Set notificationType for Incoming Call only when
InCallUi is not showing else we end up displaying HUN
even when there is a user input to accept Low Battery
MT Video call as Video.
Change-Id: Ic7d31a488bfa7bc53a436003e46c5305c730e0c1

IMS-VT: Change in position and color of Alert Dialog action buttons.

Swaps the position of action buttons in low battery alert dialog and
changes the color.
Change-Id: I8d0eb20c3ca45c6ba9182c23e8bb2229788b8343

IMS-VT: Low battery handling for calls that can change to VT

1. For outgoing upgrade requests, seek user confirmation and
   as per user input, place the outgoing upgrade request
2. For incoming upgrade requests, if user tries to accept as
   Video, seek user confirmation and accordingly handle the
   incoming upgrade request.

Change-Id: I8ea12a81b70e8e08a113d3c27721b17da65f0926
CRs-Fixed: 1012524